refactor(devti): improve chatbot configuration and prompt management #371
- Extract chatbot configuration logic into a separate dialog (McpLlmConfigDialog)
- Add prompt text management to chatbot configuration
- Refactor McpPreviewEditor to use the new dialog for configuration- Update McpToolDetailPanel with new dimensions for tool cards
